Don't segfault on events for unknown wds
Making the DirInfo struct separate meant that node_for_wd tried to
dereference a null pointer in the event that the wd was not in the
hash, so it now just returns NULL in that case.

Also make the rest of do_inotify() more robust against events for an
unknown wd other than IN_IGNORED. There should never be any so we just
print an error and exit.