Fix race condition between watching and indexing
Previously, if a creation or deletion occurred between adding the
inotify watcher and indexing the directory, jfindd would quit with
SIGABRT because an assertion would fail.

There are all sorts of complicated ways to deal with this, but the
pragmatic solution is just to ignore any events which don't seem to
make sense as they probably occurred during the race window.

Also, fix a memory leak whereby the value returned by treenode_name
was not freed in _inotify_create.